Renaming (#7)
* Application renamed

* Workflow updated

* Change made
