+++ .github/workflows/deploy.yml
@@ -4,6 +4,7 @@ on:
   push:
     branches:
       - main
+      - renaming
 
 jobs:
   deploy:
@@ -36,8 +37,8 @@ jobs:
 
       - name: Build and push image
         run: |
-          docker build --platform linux/amd64 -t ${{ secrets.REGISTRY_LOGIN_SERVER }}/${{ secrets.REGISTRY_REPOSITORY_NAME }}:${{ github.sha }} .
-          docker push ${{ secrets.REGISTRY_LOGIN_SERVER }}/${{ secrets.REGISTRY_REPOSITORY_NAME }}:${{ github.sha }}
+          docker build --platform linux/amd64 -t ${{ secrets.REGISTRY_LOGIN_SERVER }}/backend:initial .
+          docker push ${{ secrets.REGISTRY_LOGIN_SERVER }}/backend:initial
 
       - name: Build and deploy Container App
         uses: azure/container-apps-deploy-action@v1
@@ -45,4 +46,4 @@ jobs:
           acrName: ${{ secrets.AZURE_CONTAINER_REGISTRY_NAME }}
           containerAppName: ${{ secrets.AZURE_CONTAINER_APP_NAME }}
           resourceGroup: ${{ secrets.AZURE_RESOURCE_GROUP }}
-          imageToDeploy: ${{ secrets.REGISTRY_LOGIN_SERVER }}/${{ secrets.REGISTRY_REPOSITORY_NAME }}:${{ github.sha }}
+          imageToDeploy: ${{ secrets.REGISTRY_LOGIN_SERVER }}/backend:initial
