+++ Dockerfile
@@ -1,5 +1,5 @@
 FROM eclipse-temurin:21-jdk
 
-COPY build/libs/albiol-0.0.1-SNAPSHOT.jar app.jar
+COPY build/libs/higiea-0.0.1-SNAPSHOT.jar app.jar
 EXPOSE 8080
 ENTRYPOINT ["java", "-jar", "app.jar"]
