+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'albiol'
+rootProject.name = 'higiea'
