+++ src/main/java/com/ub/higiea/HigieaApplication.java
@@ -1,4 +1,4 @@
-package com.ub.albiol;
+package com.ub.higiea;
 
 import io.swagger.v3.oas.annotations.OpenAPIDefinition;
 import org.springframework.boot.SpringApplication;
@@ -6,10 +6,10 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
 @OpenAPIDefinition
-public class AlbiolApplication {
+public class HigieaApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(AlbiolApplication.class, args);
+		SpringApplication.run(HigieaApplication.class, args);
 	}
 
 }
