+++ src/main/resources/application.properties
@@ -1,7 +1,7 @@
-spring.application.name=albiol
+spring.application.name=higiea
 server.port=8080
 
-spring.r2dbc.url=${DB_URL:r2dbc:postgresql://localhost:5432/albiol_db}
+spring.r2dbc.url=${DB_URL:r2dbc:postgresql://localhost:5432/higiea_db}
 spring.r2dbc.username=${DB_USERNAME:root}
 spring.r2dbc.password=${DB_PASSWORD:root}
 spring.sql.init.mode=${SQL_INIT_MODE:always}
